v0.2.1
added:
- preview line for cue aim;
- player now can control shot intensity.

+++ Assets/Resources/Scripts/Game/PlayerBall.cs
@@ -36,7 +36,7 @@ public class PlayerBall : MonoBehaviour {
 		}
 	}
 
-	public void Shoot(float angle, int intensity) {
+	public void Shoot(float angle, float intensity) {
 		Vector2 direction = 
 			new Vector2(
 				Mathf.Cos(angle),
